Moved SplitName to io.ebean.text
