+++ src/main/java/io/ebean/text/PathProperties.java
@@ -2,7 +2,6 @@ package io.ebean.text;
 
 import io.ebean.FetchPath;
 import io.ebean.Query;
-import io.ebeaninternal.server.query.SplitName;
 
 import java.util.Collection;
 import java.util.Iterator;
