+++ src/main/java/io/ebeaninternal/api/ManyWhereJoins.java
@@ -1,9 +1,9 @@
 package io.ebeaninternal.api;
 
+import io.ebean.text.SplitName;
 import io.ebeaninternal.server.deploy.BeanProperty;
 import io.ebeaninternal.server.deploy.BeanPropertyAssocMany;
 import io.ebeaninternal.server.el.ElPropertyDeploy;
-import io.ebeaninternal.server.query.SplitName;
 import io.ebeaninternal.server.query.SqlJoinType;
 
 import java.io.Serializable;
