+++ src/main/java/io/ebeaninternal/server/deploy/TableJoin.java
@@ -1,9 +1,9 @@
 package io.ebeaninternal.server.deploy;
 
+import io.ebean.text.SplitName;
 import io.ebeaninternal.server.core.InternString;
 import io.ebeaninternal.server.deploy.meta.DeployTableJoin;
 import io.ebeaninternal.server.deploy.meta.DeployTableJoinColumn;
-import io.ebeaninternal.server.query.SplitName;
 import io.ebeaninternal.server.query.SqlJoinType;
 
 /**
