+++ src/main/java/io/ebeaninternal/server/expression/BetweenPropertyExpression.java
@@ -1,12 +1,12 @@
 package io.ebeaninternal.server.expression;
 
+import io.ebean.text.SplitName;
 import io.ebeaninternal.api.ManyWhereJoins;
 import io.ebeaninternal.api.SpiExpression;
 import io.ebeaninternal.api.SpiExpressionRequest;
 import io.ebeaninternal.api.SpiExpressionValidation;
 import io.ebeaninternal.server.deploy.BeanDescriptor;
 import io.ebeaninternal.server.el.ElPropertyDeploy;
-import io.ebeaninternal.server.query.SplitName;
 
 import java.io.IOException;
