+++ src/main/java/io/ebeaninternal/server/expression/DefaultExampleExpression.java
@@ -4,6 +4,7 @@ import io.ebean.ExampleExpression;
 import io.ebean.LikeType;
 import io.ebean.bean.EntityBean;
 import io.ebean.event.BeanQueryRequest;
+import io.ebean.text.SplitName;
 import io.ebeaninternal.api.ManyWhereJoins;
 import io.ebeaninternal.api.SpiExpression;
 import io.ebeaninternal.api.SpiExpressionRequest;
@@ -11,7 +12,6 @@ import io.ebeaninternal.api.SpiExpressionValidation;
 import io.ebeaninternal.server.deploy.BeanDescriptor;
 import io.ebeaninternal.server.deploy.BeanProperty;
 import io.ebeaninternal.server.deploy.BeanPropertyAssocOne;
-import io.ebeaninternal.server.query.SplitName;
 
 import java.io.IOException;
 import java.util.ArrayList;
