+++ src/main/java/io/ebeaninternal/server/expression/IsEmptyExpression.java
@@ -1,11 +1,11 @@
 package io.ebeaninternal.server.expression;
 
+import io.ebean.text.SplitName;
 import io.ebeaninternal.api.ManyWhereJoins;
 import io.ebeaninternal.api.SpiExpression;
 import io.ebeaninternal.api.SpiExpressionRequest;
 import io.ebeaninternal.server.deploy.BeanDescriptor;
 import io.ebeaninternal.server.el.ElPropertyValue;
-import io.ebeaninternal.server.query.SplitName;
 
 import java.io.IOException;
