+++ src/main/java/io/ebeaninternal/server/query/SqlTreeAlias.java
@@ -1,5 +1,6 @@
 package io.ebeaninternal.server.query;
 
+import io.ebean.text.SplitName;
 import io.ebeaninternal.server.deploy.BeanDescriptor;
 import io.ebeaninternal.server.el.ElPropertyDeploy;
