+++ src/main/java/io/ebeaninternal/server/query/SqlTreeNodeExtraJoin.java
@@ -2,6 +2,7 @@ package io.ebeaninternal.server.query;
 
 import io.ebean.Version;
 import io.ebean.bean.EntityBean;
+import io.ebean.text.SplitName;
 import io.ebeaninternal.api.SpiQuery;
 import io.ebeaninternal.server.deploy.BeanPropertyAssoc;
 import io.ebeaninternal.server.deploy.BeanPropertyAssocMany;
