+++ src/main/java/io/ebeaninternal/server/querydefn/OrmQueryDetail.java
@@ -1,11 +1,11 @@
 package io.ebeaninternal.server.querydefn;
 
 import io.ebean.FetchConfig;
+import io.ebean.text.SplitName;
 import io.ebeaninternal.server.deploy.BeanDescriptor;
 import io.ebeaninternal.server.deploy.BeanPropertyAssoc;
 import io.ebeaninternal.server.el.ElPropertyDeploy;
 import io.ebeaninternal.server.el.ElPropertyValue;
-import io.ebeaninternal.server.query.SplitName;
 
 import javax.persistence.PersistenceException;
 import java.io.Serializable;
